Courser R Programming Assignment No.2
 Caching the Inverse of a Matrix:
Pair of functions that are used to create a special object that
stores a matrix and caches its inverse.
First function creates a special "matrix" object that can cache its
inverse, second function computes the inverse of the special "matrix"
created by “makeCacheMatrix” above. If the inverse has already been
calculated (and the matrix has not changed), then it should retrieve
the inverse from the cache.